changed arduino values to float, instead of 16 bit integers

+++ code/IMU 9250 to serial/mpu6050_serial_to_imu/arduino/MPU9250/MPU9250.ino
@@ -108,9 +108,9 @@ void setup()
 
 void loop()
 {
-  int16_t accel_data[3];
-  int16_t gyro_data[3];
-  int16_t mag_data[3];
+  float accel_data[3];
+  float gyro_data[3];
+  float mag_data[3];
 
   // If intPin goes high, all data registers have new data
   // On interrupt, check if data ready interrupt
